+++ R/dictionary_sugar.R
@@ -1,7 +1,11 @@
 #' @title A Quick Way to Initialize Objects from Dictionaries
 #'
 #' @description
-#' Given a [Dictionary], retrieves the object with key `key`.
+#' Given a [Dictionary], retrieve objects with provided keys.
+#' * `dictionary_sugar_get()` to retrieve a single object with key `.key`.
+#' * `dictionary_sugar_mget()` to retrieve a list of objects with keys `.keys`.
+#' * `dictionary_sugar()` is deprecated in favor of `dictionary_sugar_get()`.
+#'
 #' Arguments in `...` must be named and are consumed in the following order:
 #'
 #' 1. All arguments whose names match the name of an argument of the constructor
@@ -24,8 +28,8 @@
 #' item = R6Class("Item", public = list(x = 0))
 #' d = Dictionary$new()
 #' d$add("key", item)
-#' dictionary_sugar(d, "key", x = 2)
-dictionary_sugar = function(dict, .key, ...) {
+#' dictionary_sugar_get(d, "key", x = 2)
+dictionary_sugar_get = function(dict, .key, ...) {
   assert_class(dict, "Dictionary")
   assert_string(.key)
   if (...length() == 0L) {
@@ -72,6 +76,24 @@ dictionary_sugar = function(dict, .key, ...) {
   }
 
   return(instance)
+}
+
+#' @rdname dictionary_sugar_get
+#' @export
+dictionary_sugar = dictionary_sugar_get
+
+#' @rdname dictionary_sugar_get
+#' @export
+dictionary_sugar_mget = function(dict, .keys, ...) {
+  objs = lapply(.keys, dictionary_sugar_get, dict = dict, ...)
+  if (!is.null(names(.keys))) {
+    nn = names2(.keys)
+    ii = which(!is.na(nn))
+    for (i in ii) {
+      objs[[i]]$id = nn[i]
+    }
+  }
+  objs
 }
 
 get_constructor_formals = function(x) {