+++ R/cite_bib.R
@@ -11,6 +11,8 @@
 #'   Package to read the bibtex file from.
 #' @param key (`character(1)`)\cr
 #'   Entry of the bibtex file.
+#'   If the key is `"pkg::citation"`, the [citation()] information of the
+#'   package is used instead.
 #'
 #' @return (`character(1)`) Bibentry formated as Rd.
 #'
@@ -20,29 +22,37 @@
 #' path = system.file("references.bib", package = "mlr3misc")
 #' cat(readLines(path), sep = "\n")
 #'
-#' # return value: raw rd
+#' # bibtex entry as raw Rd
 #' cite_bib("mlr3misc", "mlr")
+#'
+#' # citation info as raw Rd
+#' cite_bib("stats", "pkg::citation")
 cite_bib = function(package, key) {
   assert_string(package)
   assert_string(key)
 
-  if (!requireNamespace("bibtex", quietly = TRUE)) {
-    warningf("Could not load package 'bibtex' to parse citation '%s'", key)
-    return(sprintf("bibtex:%s", key))
+  if (key == "pkg::citation") {
+    tools::toRd(utils::citation(package))
+  } else {
+    print(key)
+    if (!requireNamespace("bibtex", quietly = TRUE)) {
+      warningf("Could not load package 'bibtex' to parse citation '%s'", key)
+      return(sprintf("bibtex:%s", key))
+    }
+
+    path = system.file("references.bib", package = package)
+    if (!file.exists(path)) {
+      warningf("Bibtex file '%s' for package '%s' does not exist", path, package)
+      return(sprintf("bibtex:%s", key))
+    }
+
+    bib = bibtex::read.bib(path)
+
+    if (key %nin% names(bib)) {
+      warningf("Key '%s' not found in references.bib of package '%s'", key, package)
+      return(sprintf("bibtex:%s", key))
+    }
+
+    paste0(tools::toRd(bib[[key]]), "\n")
   }
-
-  path = system.file("references.bib", package = package)
-  if (!file.exists(path)) {
-    warningf("Bibtex file '%s' for package '%s' does not exist", path, package)
-    return(sprintf("bibtex:%s", key))
-  }
-
-  bib = bibtex::read.bib(path)
-
-  if (key %nin% names(bib)) {
-    warningf("Key '%s' not found in references.bib of package '%s'", key, package)
-    return(sprintf("bibtex:%s", key))
-  }
-
-  paste0(tools::toRd(bib[[key]]), "\n")
 }