+++ NEWS.md
@@ -3,6 +3,8 @@
 * Fixed error in C code discovered by UBSAN checks on CRAN.
 * Added `dictionary_sugar_mget()`.
 * Renamed `dictionary_sugar()` to `dictionary_sugar_get()`.
+* Added function `cite_bib()` and Rd macro `\cite{}` to insert entries from
+  bibtex files into roxygen documentation.
 
 # mlr3misc 0.1.4